add for

+++ entities/transfer.go
@@ -3,6 +3,6 @@ package entities
 type Transfer struct {
 	ID_Pengirim       string
 	ID_penerima       string
-	Nominal           float32
+	Nominal           int
 	Tanggal_Transaksi string
 }
